Rename SearchControl StopsFinderControl
