--- src/doc/examples/ol-search.html
@@ -1,11 +1,0 @@
-<style>
-
-#mbt-search > div > div:hover {
-  background-color: lightgray;
-}
-
-</style>
-
-<div style="width: 100%; height: 100%; position: relative;">
-  <div id="map" style="width: 100%; height: 100%;"></div>
-</div>
